0.2-alpha build 1.6-11
- Spent 2 hours of my life attempting to locate a function that returned
0
- You are now able to level up
- Isn't it great!?
- If you'd like me to explain to you how the experience calculator
works, then you will be disappointed, because I am never mentioning, let
alone touching, that god-forsaken formula ever again. No really, these
calculations are the culmination of 2 consecutive hours of mashing
buttons onto an excel spreadsheet and praying to Cthulhu that it seemed
balanced.
- Oh yeah, we have an experience calculator now. Probably should have
mentioned that at the beginning...
- Sanity was lost. Someone tell me if you find it.
- Fixed Kirill's initialization of toString() method for Stats objects
by making it so experience variable was actually created rather than
being skipped. Frankly it is a miracle no one noticed that the Stats
object toString() was broken considering the Stats class is literally
used by half the classes in the game. Just saying.